Se refactoriza el código, se añaden comentarios de linea para assets, componentes. Se ordena orden de carpetas y archivos y se corrigue error de enter en Input header para evitar que busque un campo vacío.
