+++ src/routes/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
-import Home from "../pages/Home";
+import Home from "../containers/Home";
 
 const App = () => (
   <BrowserRouter>
